full-stack-web/agents/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
agno>=1.0.0
anthropic>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0
cerebras-cloud-sdk>=1.0.0
